(Phil) Fixed a simple (but serious) problem with the temp2 and temp3 readings.  (changed a u8 to a u16 in the data struct.)  Now the temp readings seem to work perfectly!
git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@136 7894878c-1315-0410-8ee3-d5d059ff63e0

+++ src/w83781d.c
@@ -213,7 +213,7 @@ struct w83781d_data {
          u8 temp;
          u8 temp_over;               /* Register value */
          u8 temp_hyst;               /* Register value */
-         u8 temp_add[2];             /* Register value */
+         u16 temp_add[2];             /* Register value */
          u16 temp_add_over[2];       /* Register value */
          u16 temp_add_hyst[2];       /* Register value */
          u8 fan_div[3];              /* Register encoding, shifted right */
